refactor(array): split array.rs types into smaller files
