+++ src/util/mod.rs
@@ -1,3 +1,4 @@
 mod cstring_scope;
 
+#[allow(unused_imports)]
 pub use cstring_scope::CStringScope;
